修改用户资料缓存方式 增加fragment dialog的binding的base类

+++ app/src/anchor/java/com/live/fox/LiveControlFragment.java
@@ -40,7 +40,7 @@ public class LiveControlFragment extends CommonLiveControlFragment implements Ch
             User user = new User();
             user.setUid(audience.getUid());
             user.setAvatar(audience.getAvatar());
-            user.setUserExp(audience.getUserExp());
+            user.setUserExp(((Double)audience.getUserExp()).floatValue());
             user.setUserLevel(audience.getUserLevel());
             if (audience.getChatHide() == 0) {
                 user.setNickname(getString(R.string.mysteriousMan));
